Add files via upload
This Python script is hosted on an EC2 instance. It queries the GCS FHIR store for resources that are selected by the user. It directs the resources to an S3 bucket. Upon storing them in the S3 bucket, it triggers a Lambda function that will extract data points from the resource for the rules engine.