Add files via upload
This lambda function is triggered when a new object is uploaded to an S3 bucket. It reads the new object (which will contain data points for certain queries, eg: Body Height in cm, Total cholesterol level, etc. Upon reading this data point, it will perform a series of arbitrary calculations on it and write the results to a destination S3 bucket.